rag/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
Authlib==1.2.1
blinker==1.6.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cryptography==41.0.4
Flask==3.0.0
frozenlist==1.4.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.28.1
pycparser==2.21
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1
urllib3==2.0.7
validators==0.22.0
weaviate-client==3.24.2
Werkzeug==3.0.0
yarl==1.9.2
